Added averaging of distance samples

+++ Lib/LidarLite.h
@@ -61,7 +61,7 @@ private:
 public:
 	LidarLite(void);
 
-	uint32_t getDistRaw(void);
+	float getDistRaw(void);
 	float getDistIn(void);
 };
